+++ packages/client/src/global.d.ts
@@ -1,0 +1,1 @@
+declare const __INTERNAL_CLIENT_VERSION__: string;