+++ packages/client/vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
 import dts from "vite-plugin-dts";
 import path from "path";
+import packageJson from "./package.json";
 
 export default defineConfig({
   plugins: [
@@ -11,6 +12,9 @@ export default defineConfig({
       outDir: "dist",
     }),
   ],
+  define: {
+    __INTERNAL_CLIENT_VERSION__: `"${packageJson.version}"`,
+  },
   build: {
     lib: {
       entry: path.resolve(__dirname, "src/index.ts"),